feat(sqlite): add uor_statement and dynamic schema
Implements the uor_statement spec in database initialization.
The database schema can be dynamically extended by uploading
arbitrary JSON-Schema to the server. The arbitrary JSON-Schema
is added to the database as a separate logical schema and then
linked to the initialized schema as nested objects under
designated anchors within the uor_statement spec.